repo/requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.7.1
Flask-WTF==1.2.2
Flask-Limiter==3.12
bleach==6.2.0
python-dotenv==1.1.0
gunicorn==23.0.0
APScheduler==3.11.0
Werkzeug==3.1.3
pytest==8.3.5
feedparser==6.0.11
cryptography
